+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # Watchman
 
 * *[v0.2.0-rc1]*
-  * Add new email template for payment receipt
+  * [FEATURE]: Added functionality to campaign_email action to send receipt emails for `paymentMade` event
 
 ## v0.1.0 (April 4, 2016)
 * *[v0.1.0-rc1]*